+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New check for copyright headers in the Makefile
 - Refactored the `main.py` module so that it's broken into smaller, more-manageable pieces
 - A page in the docs explaining the `feature_demo` example
+- Unit tests for the `spec/` folder
 
 ### Changed
 - The `merlin config` command: